nginx-ctl: print listen/control addresses clearly
The control address was being described as "listening", which confused
at least one customer, as well as Joe and Sam. This was partly due to a
bug in which app was assumed to have a `listening` event, but it does
not.

Also, removed explorer when not running in debug mode.